Fix db stuff

+++ tariff-management-service/app/src/main/resources/db/migration/V1.0.0__init.sql
@@ -2,7 +2,7 @@ CREATE SCHEMA IF NOT EXISTS tariff_management;
 
 CREATE TABLE tariff_management.tariffs
 (
-    id           BIGINT NOT NULL PRIMARY KEY,
+    id           SERIAL NOT NULL PRIMARY KEY,
     name         VARCHAR(30) NOT NULL,
     price        NUMERIC(6, 2),
     washing_time BIGINT
